Completing upgrade to LSP4Jakarta 0.2.0. Unifies completion support between MP/Jakarta. Cleaned up code, removing obsolete and/or redundant fields/methods.
