+++ src/main/java/io/openliberty/tools/intellij/lsp4jakarta/lsp/JakartaLanguageClient.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019, 2023 Red Hat, Inc. and others
+ * Copyright (c) 2019, 2024 Red Hat, Inc. and others
  * Distributed under license by Red Hat, Inc. All rights reserved.
  * This program is made available under the terms of the
  * Eclipse Public License v2.0 which accompanies this distribution,
@@ -24,6 +24,7 @@ import io.openliberty.tools.intellij.lsp4mp4ij.psi.core.utils.IPsiUtils;
 import io.openliberty.tools.intellij.lsp4mp4ij.psi.internal.core.ls.PsiUtilsLSImpl;
 import org.apache.commons.lang3.tuple.Pair;
 import org.eclipse.lsp4j.CodeAction;
+import org.eclipse.lsp4j.CompletionList;
 import org.eclipse.lsp4j.PublishDiagnosticsParams;
 import org.eclipse.lsp4jakarta.commons.*;
 import org.eclipse.lsp4jakarta.ls.api.JakartaLanguageClientAPI;
@@ -34,8 +35,6 @@ import org.eclipse.lsp4mp.commons.codeaction.CodeActionResolveData;
 import org.eclipse.lsp4mp.commons.utils.JSONUtility;
 import org.microshed.lsp4ij.client.CoalesceByKey;
 import org.microshed.lsp4ij.client.IndexAwareLanguageClient;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -45,54 +44,17 @@ import java.util.concurrent.CompletableFuture;
  * Adapted from https://github.com/redhat-developer/intellij-quarkus/blob/2585eb422beeb69631076d2c39196d6eca2f5f2e/src/main/java/com/redhat/devtools/intellij/quarkus/lsp/QuarkusLanguageClient.java
  * to match LSP4MP, Language Server for MicroProfile
  */
-public class JakartaLanguageClient extends IndexAwareLanguageClient implements JakartaLanguageClientAPI, MicroProfileProjectService.Listener {
-  private static final Logger LOGGER = LoggerFactory.getLogger(JakartaLanguageClient.class);
+public final class JakartaLanguageClient extends IndexAwareLanguageClient implements JakartaLanguageClientAPI, MicroProfileProjectService.Listener {
 
   public JakartaLanguageClient(Project project) {
     super(project);
   }
 
-  // Support the message "jakarta/java/classpath"
-  /** public CompletableFuture<List<String>> getContextBasedFilter(JakartaClasspathParams classpathParams) {
-    String uri = classpathParams.getUri();
-    List<String> snippetContexts = classpathParams.getSnippetCtx();
-    Project project = getProject();
-    var coalesceBy = new CoalesceByKey("jakarta/java/classpath", uri, snippetContexts);
-    return runAsBackground("Computing Jakarta context",
-            monitor -> PropertiesManagerForJakarta.getInstance().getExistingContextsFromClassPath(uri, snippetContexts, project), coalesceBy);
-  } **/
-
-  // Support the message "jakarta/java/cursorcontext"
-  /** public CompletableFuture<JavaCursorContextResult> getJavaCursorContext(JakartaJavaCompletionParams params) {
-    var coalesceBy = new CoalesceByKey("jakarta/java/cursorcontext", params.getUri(), params.getPosition());
-    return runAsBackground("Computing Java cursor context",
-            monitor -> {
-              IPsiUtils utils = PsiUtilsLSImpl.getInstance(getProject());
-              return PropertiesManagerForJakarta.getInstance().javaCursorContext(params, utils);
-            }, coalesceBy);
-  } **/
-
-  // Support the message "jakarta/java/diagnostics"
-  /** public CompletableFuture<List<PublishDiagnosticsParams>> getJavaDiagnostics(JakartaDiagnosticsParams jakartaParams) {
-    var coalesceBy = new CoalesceByKey("jakarta/java/diagnostics", jakartaParams.getUris());
-    IPsiUtils utils = PsiUtilsLSImpl.getInstance(getProject());
-    return runAsBackground("Computing Jakarta Java diagnostics",
-            monitor -> PropertiesManagerForJakarta.getInstance().diagnostics(jakartaParams, utils), coalesceBy);
-  } **/
-
-  // Support the message "jakarta/java/codeaction
-  /** public CompletableFuture<List<CodeAction>> getCodeAction(JakartaJavaCodeActionParams params) {
-    IPsiUtils utils = PsiUtilsLSImpl.getInstance(getProject());
-    var coalesceBy = new CoalesceByKey("jakarta/java/codeAction", params.getUri());
-    return runAsBackground("Computing Jakarta code actions",
-            monitor -> (List<CodeAction>) PropertiesManagerForJakarta.getInstance().getCodeAction(params, utils), coalesceBy);
-  } **/
-
   // Support the message "jakarta/java/diagnostics"
   @Override
   public CompletableFuture<List<PublishDiagnosticsParams>> getJavaDiagnostics(JakartaJavaDiagnosticsParams jakartaJavaDiagnosticsParams) {
-    var coalesceBy = new CoalesceByKey("jakarta/java/diagnostics", jakartaJavaDiagnosticsParams.getUris());
-    IPsiUtils utils = PsiUtilsLSImpl.getInstance(getProject());
+    final IPsiUtils utils = PsiUtilsLSImpl.getInstance(getProject());
+    final var coalesceBy = new CoalesceByKey("jakarta/java/diagnostics", jakartaJavaDiagnosticsParams.getUris());
     return runAsBackground("Computing Jakarta Java diagnostics",
             monitor -> PropertiesManagerForJakarta.getInstance().diagnostics(jakartaJavaDiagnosticsParams, utils), coalesceBy);
   }
@@ -101,7 +63,7 @@ public class JakartaLanguageClient extends IndexAwareLanguageClient implements J
   @Override
   public CompletableFuture<List<CodeAction>> getJavaCodeAction(JakartaJavaCodeActionParams jakartaJavaCodeActionParams) {
     final IPsiUtils utils = PsiUtilsLSImpl.getInstance(getProject());
-    var coalesceBy = new CoalesceByKey("jakarta/java/codeAction", jakartaJavaCodeActionParams.getUri());
+    final var coalesceBy = new CoalesceByKey("jakarta/java/codeAction", jakartaJavaCodeActionParams.getUri());
     return runAsBackground("Computing Jakarta code actions",
             monitor -> (List<CodeAction>) PropertiesManagerForJakarta.getInstance().getCodeAction(jakartaJavaCodeActionParams, utils), coalesceBy);
   }
@@ -110,24 +72,31 @@ public class JakartaLanguageClient extends IndexAwareLanguageClient implements J
   @Override
   public CompletableFuture<CodeAction> resolveCodeAction(CodeAction codeAction) {
     final IPsiUtils utils = PsiUtilsLSImpl.getInstance(getProject());
-    var coalesceBy = new CoalesceByKey("jakarta/java/resolveCodeAction");
+    final var coalesceBy = new CoalesceByKey("jakarta/java/resolveCodeAction");
     return runAsBackground("Computing Java resolve code actions", monitor -> {
-      CodeActionResolveData data = JSONUtility.toModel(codeAction.getData(), CodeActionResolveData.class);
+      final CodeActionResolveData data = JSONUtility.toModel(codeAction.getData(), CodeActionResolveData.class);
       codeAction.setData(data);
       return PropertiesManagerForJakarta.getInstance().resolveCodeAction(codeAction, utils);
     }, coalesceBy);
   }
 
+  // Support the message "jakarta/java/completion"
   @Override
   public CompletableFuture<JakartaJavaCompletionResult> getJavaCompletion(JakartaJavaCompletionParams jakartaJavaCompletionParams) {
-    return null;
+    final IPsiUtils utils = PsiUtilsLSImpl.getInstance(getProject());
+    final var coalesceBy = new CoalesceByKey("jakarta/java/completion", jakartaJavaCompletionParams.getUri(), jakartaJavaCompletionParams.getPosition());
+    return runAsBackground("Computing Java completion", monitor -> {
+      final CompletionList completionList = PropertiesManagerForJakarta.getInstance().completion(jakartaJavaCompletionParams, utils);
+      final JavaCursorContextResult cursorContext = PropertiesManagerForJakarta.getInstance().javaCursorContext(jakartaJavaCompletionParams, utils);
+      return new JakartaJavaCompletionResult(completionList, cursorContext);
+    }, coalesceBy);
   }
 
   // Support the message "jakarta/java/projectLabels"
   @Override
   public CompletableFuture<ProjectLabelInfoEntry> getJavaProjectLabels(JakartaJavaProjectLabelsParams jakartaJavaProjectLabelsParams) {
     final IPsiUtils utils = PsiUtilsLSImpl.getInstance(getProject());
-    var coalesceBy = new CoalesceByKey("jakarta/java/projectLabels",
+    final var coalesceBy = new CoalesceByKey("jakarta/java/projectLabels",
             jakartaJavaProjectLabelsParams.getUri(), jakartaJavaProjectLabelsParams.getTypes());
     return runAsBackground("Computing Java projects labels",
             monitor -> adapt(ProjectLabelManager.getInstance().getProjectLabelInfo(adapt(jakartaJavaProjectLabelsParams), utils)), coalesceBy);
@@ -137,7 +106,7 @@ public class JakartaLanguageClient extends IndexAwareLanguageClient implements J
   @Override
   public CompletableFuture<List<ProjectLabelInfoEntry>> getAllJavaProjectLabels() {
     final IPsiUtils utils = PsiUtilsLSImpl.getInstance(getProject());
-    var coalesceBy = new CoalesceByKey("jakarta/java/workspaceLabels");
+    final var coalesceBy = new CoalesceByKey("jakarta/java/workspaceLabels");
     return runAsBackground("Computing All Java projects labels",
             monitor -> adapt(ProjectLabelManager.getInstance().getProjectLabelInfo(utils)), coalesceBy);
   }
@@ -146,7 +115,7 @@ public class JakartaLanguageClient extends IndexAwareLanguageClient implements J
   @Override
   public CompletableFuture<JakartaJavaFileInfo> getJavaFileInfo(JakartaJavaFileInfoParams jakartaJavaFileInfoParams) {
     final IPsiUtils utils = PsiUtilsLSImpl.getInstance(getProject());
-    var coalesceBy = new CoalesceByKey("jakarta/java/fileInfo", jakartaJavaFileInfoParams.getUri());
+    final var coalesceBy = new CoalesceByKey("jakarta/java/fileInfo", jakartaJavaFileInfoParams.getUri());
     return runAsBackground("Computing Java file info",
             monitor -> adapt(PropertiesManagerForJava.getInstance().fileInfo(adapt(jakartaJavaFileInfoParams), utils)), coalesceBy);
   }
@@ -158,7 +127,6 @@ public class JakartaLanguageClient extends IndexAwareLanguageClient implements J
 
   @Override
   public void sourceUpdated(List<Pair<Module, VirtualFile>> sources) {
-    int i = 0;
     // not needed for Jakarta LS
   }
