+++ src/main/java/io/openliberty/tools/intellij/lsp4mp4ij/psi/core/java/diagnostics/IJavaDiagnosticsParticipant.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright (c) 2020, 2023 Red Hat Inc. and others.
+* Copyright (c) 2020, 2024 Red Hat Inc. and others.
 * All rights reserved. This program and the accompanying materials
 * which accompanies this distribution, and is available at
 * https://www.eclipse.org/legal/epl-v20.html
@@ -11,10 +11,9 @@
 *******************************************************************************/
 package io.openliberty.tools.intellij.lsp4mp4ij.psi.core.java.diagnostics;
 
-import java.util.List;
-
-import com.intellij.openapi.extensions.ExtensionPointName;
 import org.eclipse.lsp4j.Diagnostic;
+
+import java.util.List;
 
 /**
  * Java diagnostics participants API.
@@ -24,8 +23,6 @@ import org.eclipse.lsp4j.Diagnostic;
  *
  */
 public interface IJavaDiagnosticsParticipant {
-
-	public static final ExtensionPointName<IJavaDiagnosticsParticipant> EP_NAME = ExtensionPointName.create("com.redhat.devtools.intellij.quarkus.javaDiagnosticsParticipant");
 
 	/**
 	 * Returns true if diagnostics must be collected for the given context and false
